20170616_0218_VK Fixed Boot and button bugs
A. If the TSTAT sense was HIGH at boot time the ESP32 would not boot. I moved it to GPIO2, problem solved. Although now I have to disconnect TSTAT sense to burn new code on the ESP32. B.--Fixed button issue, but it still refreshes the entire web page which takes forever to load due large js files. Gotta find a work around for that next.